+++ core/callGraphInstanceRepo.go
@@ -1,0 +1,5 @@
+package core
+
+// repo for call graph instances
+type callGraphInstanceRepo interface {
+}