+++ main.go
@@ -1,21 +1,18 @@
 package main
 
 import (
-  dockercontainerengine "github.com/opspec-io/engine/pkg/containerengine/engines/docker"
   "github.com/opspec-io/engine/tcp"
   "github.com/opspec-io/engine/core"
+  containerprovider "github.com/opspec-io/engine/pkg/containerprovider/providers/fake"
+  pubsubprovider "github.com/opspec-io/engine/pkg/pubsubprovider/providers/fake"
 )
 
 func main() {
 
-  containerEngine, err := dockercontainerengine.New()
-  if (nil != err) {
-    panic(err)
-  }
-
   tcp.New(
     core.New(
-      containerEngine,
+      new(containerprovider.ContainerProvider),
+      new(pubsubprovider.PubSubProvider),
     ),
   ).Start()